implement: tutorial code review fixes - commandsHint content, localStorage try/catch, i18n TODO, extract helpers (#320)
